+++ Clinica-MVC/src/main/java/dh/backend/clinicamvc/service/impl/PacienteService.java
@@ -2,8 +2,10 @@ package dh.backend.clinicamvc.service.impl;
 
 
 import dh.backend.clinicamvc.entity.Paciente;
+import dh.backend.clinicamvc.exception.ResourcesNotFoundException;
 import dh.backend.clinicamvc.repository.IPacienteRepository;
 import dh.backend.clinicamvc.service.IPacienteService;
+import org.apache.coyote.BadRequestException;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
@@ -23,9 +25,18 @@ public class PacienteService implements IPacienteService {
     }
 
     @Override
-    public Paciente registrarPacientes(Paciente paciente){
-        LOGGER.info("Registrando : " + paciente );
-        return pacienteRepository.save(paciente);
+    public Paciente registrarPacientes(Paciente paciente) throws BadRequestException {
+
+
+        if (paciente.getNombre() != null && paciente.getApellido() != null){
+            Paciente p = pacienteRepository.save(paciente);
+            LOGGER.info("Registrando paciente : " + paciente );
+            return p;
+        }else{
+            LOGGER.info("Paciente NO (Guardado) : " + paciente );
+            throw new BadRequestException ("{\"messages\":  \"Pacienten no registrado\"}");
+        }
+
     }
 
     @Override
@@ -48,26 +59,40 @@ public class PacienteService implements IPacienteService {
     }
 
     @Override
-    public void eliminarPaciente(Integer id) {
-        LOGGER.info("Borrando Paciente con id: " + id  );
-        pacienteRepository.deleteById(id);
+    public void eliminarPaciente(Integer id) throws ResourcesNotFoundException {
+
+        Optional<Paciente> optionalPaciente = pacienteRepository.findById(id);
+        if( optionalPaciente.isPresent()){
+            LOGGER.info("Borrando Paciente con id: " + id  );
+            pacienteRepository.deleteById(id);
+        }else{
+            LOGGER.info("Borrando Paciente no fue encotrado con id: " + id  );
+            throw new ResourcesNotFoundException("{\"messages\":  \"Paciente no encontrado\"}");
+        }
+
     }
 
     /**
      * Busqueda de metodos HQL
      */
     @Override
-    public Paciente buscarPacienteporDNI(String dni) {
+    public Paciente buscarPacienteporDNI(String dni) throws BadRequestException {
         LOGGER.info("Buscar Paciente por DNI: " + dni  );
-        return pacienteRepository.findByDNI(dni);
+        Paciente p = pacienteRepository.findByDNI(dni);
+        if(p != null){
+            return pacienteRepository.findByDNI(dni);
+        }else{
+            throw new BadRequestException ("{\"messages\":  \"DNI de paciente no encontrado\"}");
+
+        }
+
     }
 
-//    @Override
-//    public List<Paciente> buscarPacienteDomicilioPorProvincia(String provincia) {
-//        LOGGER.info("Borrando Paciente por provincia: " + provincia  );
-//        pacienteRepository.findByProvincia(provincia);
-//        return null;
-//    }
+    @Override
+    public List<Paciente> buscarPacienteDomicilioPorProvincia(String provincia) {
+        LOGGER.info("Borrando Paciente por provincia: " + provincia  );
+        return pacienteRepository.findByProvincia(provincia);
+    }
 
 
 }
